splitted the code and put it in separate folders
