+++ controllers/contacts/addContact.js
@@ -1,16 +1,23 @@
-const validationSchema = require("../../schemas");
-const contactsOperations = require("../../models/contacts");
+const Joi = require("joi");
+const contactsOperations = require("../../models/contactsOperations");
+
+const joiShema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().required(),
+  phone: Joi.string().required(),
+});
 
 const addContact = async (req, res, next) => {
-  const { error } = validationSchema.validate(req.body);
-  if (error) {
-    res.json({
-      status: "error",
-      code: 400,
-      message: "missing required name field",
-    });
-  }
   try {
+    const { error } = joiShema.validate(req.body);
+    if (error) {
+      res.json({
+        status: "error",
+        code: 400,
+        message: "missing required name field",
+      });
+    }
+
     const result = await contactsOperations.addContact(req.body);
     res.json({
       status: "success",
