+++ controllers/contacts/updateContact.js
@@ -1,16 +1,23 @@
-const validationSchema = require("../../schemas");
-const contactsOperations = require("../../models/contacts");
+const Joi = require("joi");
+const contactsOperations = require("../../models/contactsOperations");
+
+const joiShema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().required(),
+  phone: Joi.string().required(),
+});
 
 const updateContact = async (req, res, next) => {
-  const { error } = validationSchema.validate(req.body);
-  if (error) {
-    res.json({
-      status: "error",
-      code: 400,
-      message: "missing required name field",
-    });
-  }
   try {
+    const { error } = joiShema.validate(req.body);
+    if (error) {
+      res.json({
+        status: "error",
+        code: 400,
+        message: "missing required name field",
+      });
+    }
+
     const { contactId } = req.params;
     const result = await contactsOperations.updateContact(contactId, req.body);
