+++ models/contactsOperations/dbPath.js
@@ -1,0 +1,5 @@
+const path = require("path");
+
+const dbPath = path.join(__dirname, "db.json");
+
+module.exports = dbPath;
