+++ models/contactsOperations/updateContact.js
@@ -1,5 +1,5 @@
 const fs = require("fs/promises");
-const contactsPath = require("../contacts.json");
+const contactsDatabase = require("./dbPath");
 const listContacts = require("./listContacts");
 
 const updateContact = async (contactId, body) => {
@@ -7,7 +7,7 @@ const updateContact = async (contactId, body) => {
   const updateId = contacts.findIndex((item) => item.id === contactId);
 
   contacts[updateId] = { id: contactId, ...body };
-  await fs.writeFile(contactsPath, JSON.stringify(contacts, null, 2));
+  await fs.writeFile(contactsDatabase, JSON.stringify(contacts, null, 2));
   return contacts[updateId];
 };
 module.exports = updateContact;
