--- schemas/index.js
@@ -1,4 +1,0 @@
-const validationSchema = require("./validationSchema");
-module.exports = {
-  validationSchema,
-};
