--- schemas/validationSchema.js
@@ -1,9 +1,0 @@
-const Joi = require("joi");
-
-const validationSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-});
-
-module.exports = validationSchema;
